小地圖相關 怪物設定相關
終點如何算
需要條件
結算畫面

怪物設定：
1.將血量拿掉
統一吃士氣值

2.修改部位血量
更名為耐久度
功能跟之前一樣
只是換名子

3.怪物新增設定獲得積分欄位

小地圖設定：
1.新增蚊子icon onminimap
影文字框給他
靠北內容依照士氣值
士氣值會照時間做1/10的變化
(原生蚊還沒連動)

2.新增結算畫面
有興修分數跟回到大地圖的按鈕

3.倘若怪物死亡
或者原生蚊死光
小地圖上都會做圖片變化

+++ P_MQ_2016 0908-0909Digital Taipei/Assets/06_GameScene/Ver3/Scripts/OnCameraForShootMQ.cs
@@ -114,7 +114,7 @@ public class OnCameraForShootMQ : MonoBehaviour
             myautobulletcount_text.text = myAutoFireBulletAmount.ToString();
             myautobulletbar_image.fillAmount = (float)myAutoFireBulletAmount / (float)myAutoFireBulletFullAmount;
 
-            if (myAutoFireTimer > 1/ (float)GameObject.Find("Canvas").GetComponent<onCanvasForUIControll>().myLocalMQ_CreateSpeed) {
+            if (myAutoFireTimer > 1/ GameObject.Find("Canvas").GetComponent<onCanvasForUIControll>().myLocalMQ_CreateSpeed) {
                 if (GameObject.Find("Canvas").GetComponent<onCanvasForUIControll>().myLocalMQ_Amount <= 0) { }
                 else {
                     GameObject.Find("Canvas").GetComponent<onCanvasForUIControll>().myLocalMQ_Amount--;
